+++ src/view/group/index/WithdrawDialog/query.ts
@@ -1,0 +1,8 @@
+import { withdrawGroup } from "@/api/group";
+import { useMutation } from "@tanstack/react-query";
+
+export const useWithdrawMutation = (groupId: number) => {
+  return useMutation({
+    mutationFn: () => withdrawGroup(groupId),
+  });
+};
